Docker/requirements.txt
pymongo==3.11.0
selenium==4.19.0 
apache-airflow-providers-mongo==4.0.0
# apache-airflow==2.9.0
# webdriver
python-dotenv==1.0.1 
googlemaps==4.10.0
requests==2.31.0
boto3==1.34.86